Fix handling of path to /cgit/
As a result of the fix to Python issue 19435 [0], the `run_cgi()`
function in CGIHTTPRequestHandler no longer uses `self.path` in order to
calculate the script directory, but instead uses the second component of
`self.cgi_info`.

Since CGit expects to see the complete URL, change the use of
`self.cgi_info` to put the complete path in the second component.  Also
update `translate_path()` to handle this.

[0] http://bugs.python.org/issue19435

Reported-by: Dennis Kaarsemaker <dennis@kaarsemaker.net>
Signed-off-by: John Keeping <john@keeping.me.uk>