Fix handling of path to /cgit/ (again)
Python issue 21323 [0] fixed a regression caused by the previous fix to
issue 19435 [1] affecting CGI scripts in subdirectories by changing the
way in which the path is calculated.  However, this means that if the
path starts with '/' it is now split so that $SCRIPT_NAME is empty,
which results in CGit using the entire PATH_INFO to find the repository.

Fix this by removing the leading '/' in cgi_info (if there is one) so
that the split gives $SCRIPT_NAME as "cgit", which allows CGit to
calculate its virtual root correctly.

[We fixed issues caused by the original upstream commit in commit
 aa273b4 (Fix handling of path to /cgit/, 2014-07-27).]

[0] http://bugs.python.org/issue21323
[1] http://bugs.python.org/issue19435

Signed-off-by: John Keeping <john@keeping.me.uk>